feat: add utility function for scaled icons and update icon handling in LSP server and Bolt file type
